requirements.txt
pywinauto==0.6.8
pillow==10.1.0
jsonschema==4.20.0
colorlog==6.8.0
